Fix memory leak caused by monitoring app switching
The code that monitors app switching, relinquishing control
of the media keys to another process in its white list, was
causing a huge memory leak. Instead of fixing it, I decided
it wasn't a feature I wanted to support. This change simply
comments out the feature and fixes the memory problem.